fix(shellcheck): Use json1 format for correct column reporting
The currently used `json` format counts hard tabs as 8 columns, when
reporting column offsets. This leads to misaligned error display.

The `json1` format counts tabs as 1 column. This format outputs an
object with the errors array under the "comments" key.

The new format was added in shellcheck
[v0.7.0](https://github.com/koalaman/shellcheck/blob/master/CHANGELOG.md#v070---2019-07-28)

About the output format:
https://github.com/koalaman/shellcheck/wiki/Integration#json-output